+++ src/mission/imap/+mag/+imap/+hk/Status.m
@@ -2,6 +2,8 @@ classdef Status < mag.HK
 % STATUS Class containing MAG status HK packet data.
 
     properties (Dependent)
+        % CPUILDE CPU idle percentage.
+        CPUIdle (:, 1) double
         % FOBACTIVE Outboard sensor active.
         FOBActive (:, 1) logical
         % FIBACTIVE Inboard sensor active.
@@ -9,6 +11,10 @@ classdef Status < mag.HK
     end
 
     methods
+
+        function cpuIdle = get.CPUIdle(this)
+            cpuIdle = this.Data.CPUIDLE;
+        end
 
         function fobActive = get.FOBActive(this)
             fobActive = this.Data.FOBSTAT;